pyproject.toml
[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "better-word-slack"
version = "0.1"
description = "Improve your Slack experience, every little bit helps"
readme = "README.md"
authors = [
    { name = "Vincent Wang", email = "awayfromtty@mcpnow.io" }
]
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["mcp", "phabricator", "phorge", "api", "client", "server", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Bug Tracking"
]
dependencies = [
    "slack_sdk",
    "python-dotenv",
    "litellm",
]

[project.urls]
Homepage = "https://github.com/mcpnow-io/better-word-slack"
"Bug Reports" = "https://github.com/mcpnow-io/better-word-slack/issues"
Source = "https://github.com/mcpnow-io/better-word-slack"
Wiki = "https://github.com/mcpnow-io/better-word-slack/wiki"

[project.optional-dependencies]
dev = [
    "flake8",
    "pre-commit",
    "pytest",
    "pytest-cov",
]
gemini = [
    "google-genai",
]

[project.scripts]
better-word = "src.better_word:main"

[tool.setuptools]
packages = ["src"]
include-package-data = true
zip-safe = false
